backend/requirements.txt
# Screenshot Studio - Python Dependencies

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Security middleware (B2, B3)
Flask-Cors==4.0.0
Flask-Limiter==3.5.0

# Env loading (B8 — keeps secrets out of config.py)
python-dotenv==1.0.0

# Screenshot Engine
playwright==1.40.0

# Image Processing
Pillow==10.1.0
PyMuPDF==1.23.26  # PDF conversion to images
requests==2.31.0

# Production servers (pick one based on platform)
gunicorn==21.2.0  # Linux/Mac
waitress==2.1.2   # Windows-friendly WSGI server

# PowerPoint Automation (Windows only)
python-pptx==0.6.23  # PowerPoint file manipulation
pywin32==306; sys_platform == 'win32'  # Windows COM automation for video export

# Testing
hypothesis==6.92.1  # Property-based testing
